Overhaul data versioning system.
The old system stored data in columns of composite type, which, it
turns out, do not support foreign key constraints on their constituent
fields.  The new system stores data in columns directly, which gives
rise to a difficulty: how can we transfer data between foo_v and foo
without knowing the field names?  This is possible if the transferring
command is an INSERT, using a trick; see the trigger functions in
versionize() for details.  In order that every such transfer be
an INSERT, the triggers must transfer from foo to foo_v (not the
reverse as in the old system), so the user should update foo directly
as normal.  For deletions, then, we don't get the correct transaction
ID from the record itself, so we look it up in the transaction table;
for consistency, we do the same for inserts and updates as well.